Long converted, DTO development, validator created

+++ src/main/java/pl/gr/veterinaryapp/controller/rest/PetRestController.java
@@ -31,12 +31,12 @@ public class PetRestController {
     private final PetMapper mapper;
 
     @DeleteMapping(path = "/{id}")
-    public void delete(@PathVariable int id) {
+    public void delete(@PathVariable Long id) {
         petService.deletePet(id);
     }
 
     @GetMapping(path = "/{id}", produces = MediaTypes.HAL_JSON_VALUE)
-    public PetResponseDto getPet(@AuthenticationPrincipal User user, @PathVariable long id) {
+    public PetResponseDto getPet(@AuthenticationPrincipal User user, @PathVariable Long id) {
         var pet = mapper.map(petService.getPetById(user, id));
         addLinks(pet);
         return pet;
@@ -65,12 +65,12 @@ public class PetRestController {
         return pet;
     }
 
-    public Link createClientLink(long id) {
+    public Link createClientLink(Long id) {
         return linkTo(methodOn(ClientRestController.class).getClient(id))
                 .withRel("client");
     }
 
-    public Link createAnimalLink(long id) {
+    public Link createAnimalLink(Long id) {
         return linkTo(methodOn(AnimalRestController.class).getAnimal(id))
                 .withRel("animal");
     }
